Fix max lim

+++ eims/src/main/java/com/example/eims/dto/cost/CostDetailDTO.java
@@ -16,6 +16,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -25,8 +26,8 @@ public class CostDetailDTO {
     private Long userId;
     private Long facilityId;
     private String costItem;
-    private Float costAmount;
-    private Float paidAmount;
+    private BigDecimal costAmount;
+    private BigDecimal paidAmount;
     private String issueDate;
     private String note;
     private boolean status;
