+++ eims/src/main/java/com/example/eims/dto/payroll/CreatePayrollDTO.java
@@ -15,6 +15,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Data
@@ -24,7 +25,7 @@ public class CreatePayrollDTO {
     private Long employeeId;
     private String phone;
     private String payrollItem;
-    private Float payrollAmount;
+    private BigDecimal payrollAmount;
     private String issueDate;
     private String note;
     private boolean status;
